cloud-run/education-streamlit/requirements.txt
streamlit
google-cloud-aiplatform
google-cloud-logging
google-cloud-speech
google-cloud-storage
pytesseract 
PyPDF2
textract
